Task information loaded in smaller portions for speed
In order to support thread reporting commands as 'info threads' or
'thread apply all bt', as well as putting GDB in a proper thread on breakpoint
or other halt reason for unwinding the proper stack, gdb proxy has to support
some of the thread reporting requests:
 - 'qfThreadInfo'
 - 'qsThreadInfo'
 - 'qThreadExtraInfo'

Previously it was done exhaustively - each time the a core halts for any
reason - complete block of thread information has been updated.

This has turned out to be too time consuming, which drastically affects
debug experience in a negative way.

Now thread(or task) information readout from target memory is performed more
selectively, only reading information that is currently required.

So for 'info thread' command the amount of data read would be a little bit
faster, because complete 'struct task' is not read out, instead only some of
its fields are read. But this will not bring really big performance change,
however for commands like 'si', 'ni', where we debug in stepping mode and we
don't want to look on other threads, there will be much less thread info read-
out, because gdb will not ask for much.